+++ src/main/kotlin/kea/dpang/auth/redis/repository/RefreshTokenRepository.kt
@@ -3,5 +3,5 @@ package kea.dpang.auth.redis.repository
 import kea.dpang.auth.redis.entity.RefreshToken
 import org.springframework.data.repository.CrudRepository
 
-interface RefreshTokenRepository : CrudRepository<RefreshToken, Int> {
+interface RefreshTokenRepository : CrudRepository<RefreshToken, Long> {
 }
